refactor: 파일 조회 컨트롤러 위치 변경 - MetadataController -> FileController - base위치 파악 위해 FileService 에 FolderService 의존 생성 - 변경 지점에 따른 테스트 코드 수정
