+++ src/main/java/com/c4cometrue/mystorage/file/FileController.java
@@ -3,12 +3,14 @@ package com.c4cometrue.mystorage.file;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.c4cometrue.mystorage.dto.FileDeleteRequest;
 import com.c4cometrue.mystorage.dto.FileDownloadRequest;
+import com.c4cometrue.mystorage.dto.FileUploadRequest;
 
 import lombok.RequiredArgsConstructor;
 
@@ -17,6 +19,12 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class FileController {
 	private final FileService fileService;
+
+	@PostMapping("/files")
+	@ResponseStatus(HttpStatus.CREATED)
+	public void uploadFile(FileUploadRequest req) {
+		fileService.uploadFile(req.multipartFile(), req.userId(), req.parentId());
+	}
 
 	@DeleteMapping
 	@ResponseStatus(HttpStatus.NO_CONTENT)
