+++ src/test/java/com/c4cometrue/mystorage/metadata/MetadataControllerTest.java
@@ -25,13 +25,6 @@ class MetadataControllerTest {
 	private MetadataService metadataService;
 
 	@Test
-	@DisplayName("업로드 파일 테스트")
-	void uploadFileTest() {
-		metadataController.uploadFile(FileUploadRequest.of(MOCK_MULTIPART_FILE, USER_ID, PARENT_ID));
-		verify(metadataService, times(1)).uploadFile(MOCK_MULTIPART_FILE, USER_ID, PARENT_ID);
-	}
-
-	@Test
 	@DisplayName("특정 폴더 상세 조회 테스트")
 	void getFolderContentsTest() {
 		given(metadataService.getFolderContents(FOLDER_ID, USER_ID)).willReturn(eq(anyList()));
